backend/requirements.txt
# Flask
# Flask-Cors
# tensorflow
# numpy
# Pillow
# requests

Flask
Flask-Cors
google-generativeai
python-dotenv
Pillow
requests